Restructure how parts, chapters, sections work
Scribble's section hierarchy assumes that all organization
is done using sections/parts. So it's best to actually use the
`section` function to build chapters since it cooperates with
`include-section` in the right way.

The previous implementation of chapters would screw up the
section numbering when a chapter or part is included.

However, parts do need their own includer since the part style
has to be applied specially. This commits adds an `include-part`
form.

This is a backwards incompatible change. Updating old code is not hard
though.

Sections below chapters should be switched to `subsection` (and
repeat for lower sections). Including parts requires `include-part`.
Included chapters and parts should use `title`.